+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.2.0 IN PROGRESS
   * SI-172 Add Duplicate of Sequences for number generator sequences in settings
   * SI-151 Enable dynamic ${current_year} token in number generator output templates
+  * SI-173 Duplicate sequence: Show callout if using non unique code
 
 ## 4.1.1 2026-05-28
   * Updated dependencies for Trillium release